+++ pyrigor/checkers/__init__.py
@@ -8,6 +8,7 @@ from pyrigor.checkers.pyr401_namedtuple_returns import find_violations as _pyr40
 from pyrigor.checkers.pyr402_keyword_only_arguments import find_violations as _pyr402
 from pyrigor.checkers.pyr403_keyword_only_single_argument import find_violations as _pyr403
 from pyrigor.checkers.pyr405_namedtuple_parameters import find_violations as _pyr405
+from pyrigor.checkers.pyr406_return_values_used import find_violations as _pyr406
 from pyrigor.rules import Rule
 from pyrigor.violations import Violation
 
@@ -29,12 +30,14 @@ class RegisteredChecker(NamedTuple):
     find_violations: _CheckerFun
 
 
+# noinspection PyTypeChecker
 CHECKERS: tuple[RegisteredChecker, ...] = (
     RegisteredChecker(rule=Rule.PYR301, find_violations=_pyr301),
     RegisteredChecker(rule=Rule.PYR401, find_violations=_pyr401),
     RegisteredChecker(rule=Rule.PYR402, find_violations=_pyr402),
     RegisteredChecker(rule=Rule.PYR403, find_violations=_pyr403),
     RegisteredChecker(rule=Rule.PYR405, find_violations=_pyr405),
+    RegisteredChecker(rule=Rule.PYR406, find_violations=_pyr406),
 )
 
 __all__ = ["CHECKERS", "RegisteredChecker"]